+++ src/ui/main_ui.h
@@ -5,12 +5,15 @@
 #include "cx.h"
 #include "os_io_seproxyhal.h"
 
-#include "sky.h"
-#include "ux.h"
+#include "../skycoin-api/skycoin_crypto.h"
+#include "../ux.h"
 
-#include "glyphs.h"
+#include "../glyphs.h"
 
 extern ux_state_t ux;
+
+extern int current_offset;
+extern int direction;
 
 extern const ux_menu_entry_t menu_main[];
 extern const ux_menu_entry_t menu_settings[];